final commit of libft

+++ 42_Projects/42_Formation/libft/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: andrefil <andrefil@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/11 19:09:58 by andrefil          #+#    #+#             */
-/*   Updated: 2023/11/03 21:49:16 by andrefil         ###   ########.fr       */
+/*   Updated: 2023/11/05 13:00:38 by andrefil         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,11 +15,11 @@
 void	*ft_memset(void *str, int ch, size_t num)
 {
 	char	*string;
-	int		index;
+	size_t	index;
 
 	string = (char *)str;
 	index = 0;
-	while (index < (int)num)
+	while (index < num)
 		string[index++] = ch;
 	return (str);
 }
